leitura dos testes do arquivo

+++ tads/lista_processos/lista_processo.c
@@ -17,7 +17,7 @@ Celula *inicializa_celula(Celula *celula, Processo *processo)
   celula->processo = processo;
   return celula;
 }
-Lista *inicializa_lista(Lista *lista, int tamanho)
+Lista *inicializa_lista(Lista *lista, long int tamanho)
 {
   lista = (Lista *)malloc(sizeof(Lista));
   lista->tam = tamanho;
@@ -192,7 +192,7 @@ void imprime_lista(Lista *lista)
   printf("\nLista de processos:\n");
   Celula aux = lista->plista[lista->primeiro];
   if(lista->numCelOcupadas == 1){
-    printf("%d\n", aux.processo->PID);
+    printf("%llu\n", aux.processo->PID);
     return;
   }
   while (aux.prox != -1)
